problem-solver/py/requirements.txt
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
httplib2==0.31.0
idna==3.10
proto-plus==1.26.1
protobuf==6.33.0
py-sc-client==0.4.0
py-sc-kpm==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dotenv==1.2.1
requests==2.31.0
rsa==4.9.1
ruff==0.14.1
uritemplate==4.2.0
websocket-client==1.8.0
